voxelization/pyproject.toml
[project]
name = "voxelization"
version = "1.0.0"
description = "Voxelization Network for Point Cloud to Volume Conversion"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "scipy>=1.11.0",
    "tqdm>=4.65.0",
    "trimesh>=4.0.0",
    "tensorboard>=2.12.0",
    "plyfile>=1.0.0",
    "open3d>=0.18.0",
    "scikit-image>=0.21.0",
    "opencv-python>=4.8.0",
    "pykdtree>=1.3.7",
    "pynrrd>=1.0.0",
    "PyMCubes>=0.1.4",
    "diplib>=3.4.0",
    "imageio>=2.31.0",
    "plotly>=5.14.0",
    # Note: pytorch3d and pytorch-scatter need to be installed separately
    # See installation instructions in README.md
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipdb>=0.13.0",
    "pytest>=7.0.0",
]

